Add files via upload
This repository presents my AWS portfolio, including four important modules related to theory and practice. Modules 1-3 are concerned with fundamental principles of AWS, such as cloud service and deployment types, world infrastructure, billing, and account structure, with corresponding case studies and graphics. Module 4-7 encompasses hands-on labs in IAM, VPC, Lambda, and EBS with the focus on the AWS Shared Responsibility Model and cost-effective cloud architecture, as well as screenshots and labs confirmation. The GitHub repository was lastly created and presented on GitHub Pages as a live professional display of completed AWS work.